refactor: move use*Translations hooks out of src/data into src/app
These are React/i18n hooks (useTranslation), not pure domain data —
they broke the "no React imports in src/data/" rule (technical-debt.md A2).
