+++ src/app/PyramidLevel/useLootDetermination.tsx
@@ -6,7 +6,7 @@ import { determineMapPieceLoot } from "./mapPieceLogic"
 import { determineInventoryLootForCurrentRuns } from "./inventoryLootLogic"
 import { determineExpeditionBonus } from "./expeditionBonusLogic"
 import { useInventory } from "@/app/Inventory/useInventory"
-import { useInventoryItem } from "@/data/useInventoryTranslations"
+import { useInventoryItem } from "@/app/translations/useInventoryTranslations"
 import { HieroglyphTile } from "@/ui/atoms/HieroglyphTile"
 import { getItemFirstLevel } from "@/data/itemLevelLookup"
