+++ src/app/pages/Collection.tsx
@@ -3,8 +3,8 @@ import { use, useEffect, useMemo, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { Page } from "@/ui/atoms/Page"
 import { HieroglyphTile } from "@/ui/atoms/HieroglyphTile"
-import { useInventoryCategory } from "@/data/useInventoryTranslations"
-import { useTreasureCategory } from "@/data/useTreasureTranslations"
+import { useInventoryCategory } from "@/app/translations/useInventoryTranslations"
+import { useTreasureCategory } from "@/app/translations/useTreasureTranslations"
 import { getItemFirstLevel } from "@/data/itemLevelLookup"
 import { useInventory } from "@/app/Inventory/useInventory"
 import { useProgression } from "@/app/state/useProgression"
