+++ src/app/pages/TableauInventory.tsx
@@ -1,7 +1,7 @@
 import { useMemo, type FC } from "react"
 import { useJourneys, type CombinedJourneyState } from "../state/useJourneys"
 import { journeys, type TreasureTombJourney } from "@/data/journeys"
-import { useTableauTranslations } from "@/data/useTableauTranslations"
+import { useTableauTranslations } from "@/app/translations/useTableauTranslations"
 import { generateNewSeed, mulberry32 } from "@/game/random"
 import {
   buildTombCalculationSettings,
