+++ src/app/translations/useInventoryTranslations.ts
