+++ src/app/translations/useTreasureTranslations.ts
