+++ src/ui/molecules/HieroglyphUnlockPanel.tsx
@@ -1,6 +1,6 @@
 import { useState, useRef, type FC } from "react"
 import { HieroglyphTile } from "@/ui/atoms/HieroglyphTile"
-import { useTreasureItem } from "@/data/useTreasureTranslations"
+import { useTreasureItem } from "@/app/translations/useTreasureTranslations"
 import { getItemFirstLevel } from "@/data/itemLevelLookup"
 import type { Difficulty } from "@/data/difficultyLevels"
