+++ src/ui/organisms/JourneyCard.tsx
@@ -1,6 +1,6 @@
 import type { FC, PropsWithChildren } from "react"
 import clsx from "clsx"
-import type { TranslatedJourney } from "@/data/useJourneyTranslations"
+import type { TranslatedJourney } from "@/app/translations/useJourneyTranslations"
 import { DifficultyPill } from "@/ui/atoms/DifficultyPill"
 
 type JourneyCardLabels = {
